slackbot-test/frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["yongjian.tek <yongjian.tek@grabtaxi.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.4"
requests = "^2.32.3"
pillow = "^10.3.0"
requests-toolbelt = "^1.0.0"
streamlit = "^1.35.0"
pdf2image = "^1.17.0"
streamlit-pdf-viewer = "^0.0.12"
langchain = "^0.2.1"
langchain-community = "^0.2.1"
openai = "^1.31.0"
langchain-openai = "^0.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
